Fix incorrect requests to backend
Incorrect requests were sent when the month state was changed then the
force was changed afterwards. With the new force change the month picker
would display the most recent month available as selected but the data
was retrieved using the previous month saved in the state variable.
Meaning the data was incorrect for the month displayed.